+++ files/terraform/main.tf
@@ -1,1 +1,5 @@
 provider "alicloud" {}
+
+resource "null_resource" "dummy_resource" {}
+
+resource "random_uuid" "uuid" {}