+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+* 23-07-2021, version 0.0.6
+  * fix how alicloud binaries are stored
+* 23-07-2021, version 0.0.5
+  * [terraform 1.0.3](https://github.com/hashicorp/terraform/releases/tag/v1.0.3)
+  * added hashicorp/null and hashicorp/random
 * 17-07-2021, version 0.0.4
   * [alicloud v1.127.0](https://github.com/aliyun/terraform-provider-alicloud/releases/tag/v1.127.0)
 * 13-07-2021, version 0.0.3